first successful librechroot creation

+++ src/stage2/stage2.sh
@@ -41,10 +41,15 @@ check_exe pacman
 check_exe sed
 check_exe tar
 
+# required for dbus
+check_file /usr/share/aclocal/ax_append_flag.m4
+
 # prepare for the build
 . "$_srcdir"/prepare_chroot.sh
 . "$_srcdir"/prepare_makepkg.sh
 . "$_srcdir"/prepare_deptree.sh
+
+msg "starting $CARCH cross-build"
 
 # keep building packages until the deptree is empty
 while [ -s "$_deptree" ]; do
@@ -58,28 +63,34 @@ while [ -s "$_deptree" ]; do
   [ "x$_pkgarch" == "xany" ] || _pkgarch=$CARCH
 
   msg "makepkg: $_pkgname"
-  msg "  remaining pkges: $(cat "$_deptree" | wc -l)"
+  msg "  remaining packages: $(cat "$_deptree" | wc -l)"
 
-  # prepare directories
-  _pkgdir="$_makepkgdir"/$_pkgname/pkg/$_pkgname
-  rm -rf "$_makepkgdir"/$_pkgname
-  mkdir -pv "$_makepkgdir"/$_pkgname
-  pushd "$_makepkgdir"/$_pkgname >/dev/null
+  echo -n "checking for built $_pkgname package ... "
+  pacman --config "$_chrootdir"/etc/pacman.conf -r "$_chrootdir" -Syyi $_pkgname &>/dev/null \
+    && _have_pkg=yes || _have_pkg=no
+  echo $_have_pkg
 
-  if [ "x$_pkgarch" == "xany" ]; then
-    # repackage arch=(any) packages
-    _pkgver=$(pacman -Si $_pkgname | grep '^Version' | awk '{print $3}')
-    pacman -Sw --noconfirm --cachedir "$_pkgdest" $_pkgname
-    ln -s "$_pkgdest"/$_pkgname-$_pkgver-any.pkg.tar.xz \
-      "$_makepkgdir"/$_pkgname/$_pkgname-$_pkgver-any.pkg.tar.xz
-  elif [ "x$_pkgname" == "xca-certificates-mozilla" ]; then
-    # repackage ca-certificates-mozilla to avoid building nss
-    _pkgver=$(pacman -Si $_pkgname | grep '^Version' | awk '{print $3}')
-    pacman -Sw --noconfirm --cachedir . $_pkgname
-    mkdir tmp && bsdtar -C tmp -xf $_pkgname-$_pkgver-*.pkg.tar.xz
-    mkdir -p "$_pkgdir"/usr/share/
-    cp -rv tmp/usr/share/ca-certificates/ "$_pkgdir"/usr/share/
-    cat > "$_pkgdir"/.PKGINFO << EOF
+  if [ "x$_have_pkg" == "xno" ]; then
+    # prepare directories
+    _pkgdir="$_makepkgdir"/$_pkgname/pkg/$_pkgname
+    rm -rf "$_makepkgdir"/$_pkgname
+    mkdir -pv "$_makepkgdir"/$_pkgname
+    pushd "$_makepkgdir"/$_pkgname >/dev/null
+
+    if [ "x$_pkgarch" == "xany" ]; then
+      # repackage arch=(any) packages
+      _pkgver=$(pacman -Si $_pkgname | grep '^Version' | awk '{print $3}')
+      pacman -Sw --noconfirm --cachedir "$_pkgdest" $_pkgname
+      ln -s "$_pkgdest"/$_pkgname-$_pkgver-any.pkg.tar.xz \
+        "$_makepkgdir"/$_pkgname/$_pkgname-$_pkgver-any.pkg.tar.xz
+    elif [ "x$_pkgname" == "xca-certificates-mozilla" ]; then
+      # repackage ca-certificates-mozilla to avoid building nss
+      _pkgver=$(pacman -Si $_pkgname | grep '^Version' | awk '{print $3}')
+      pacman -Sw --noconfirm --cachedir . $_pkgname
+      mkdir tmp && bsdtar -C tmp -xf $_pkgname-$_pkgver-*.pkg.tar.xz
+      mkdir -p "$_pkgdir"/usr/share/
+      cp -rv tmp/usr/share/ca-certificates/ "$_pkgdir"/usr/share/
+      cat > "$_pkgdir"/.PKGINFO << EOF
 pkgname = $_pkgname
 pkgver = $_pkgver
 pkgdesc = Mozilla's set of trusted CA certificates
@@ -88,69 +99,70 @@ builddate = $(date '+%s')
 size = 0
 arch = $_pkgarch
 EOF
-    cd "$_pkgdir"
-    env LANG=C bsdtar -czf .MTREE \
-      --format=mtree \
-      --options='!all,use-set,type,uid,gid,mode,time,size,md5,sha256,link' \
-      .PKGINFO *
-    env LANG=C bsdtar -cf - .MTREE .PKGINFO * | xz -c -z - > \
-      "$_pkgdest"/$_pkgname-$_pkgver-$_pkgarch.pkg.tar.xz
-    ln -s "$_pkgdest"/$_pkgname-$_pkgver-$_pkgarch.pkg.tar.xz \
-      "$_makepkgdir"/$_pkgname/$_pkgname-$_pkgver-$_pkgarch.pkg.tar.xz
-  else
-    # acquire the pkgbuild and auxiliary files
-    _libre=https://www.parabola.nu/packages/libre/x86_64/$_pkgname/
-    _core=https://www.archlinux.org/packages/core/x86_64/$_pkgname/
-    _extra=https://www.archlinux.org/packages/extra/x86_64/$_pkgname/
-    _community=https://www.archlinux.org/packages/community/x86_64/$_pkgname/
-    for url in $_libre $_core $_extra $_community; do
-      if ! curl -s $url | grep -iq 'not found'; then
-        src=$(curl -s $url | grep -i 'source files' | cut -d'"' -f2 | sed 's#/tree/#/plain/#')
-        for link in $(curl -sL $src | grep '^  <li><a href' | cut -d"'" -f2 \
-            | sed "s#^#$(echo $src | awk -F/ '{print $3}')#"); do
-          wget -q $link -O $(basename ${link%\?*});
-        done
-        break
-      fi
-    done
+      cd "$_pkgdir"
+      env LANG=C bsdtar -czf .MTREE \
+        --format=mtree \
+        --options='!all,use-set,type,uid,gid,mode,time,size,md5,sha256,link' \
+        .PKGINFO *
+      env LANG=C bsdtar -cf - .MTREE .PKGINFO * | xz -c -z - > \
+        "$_pkgdest"/$_pkgname-$_pkgver-$_pkgarch.pkg.tar.xz
+      ln -s "$_pkgdest"/$_pkgname-$_pkgver-$_pkgarch.pkg.tar.xz \
+        "$_makepkgdir"/$_pkgname/$_pkgname-$_pkgver-$_pkgarch.pkg.tar.xz
+    else
+      # acquire the pkgbuild and auxiliary files
+      _libre=https://www.parabola.nu/packages/libre/x86_64/$_pkgname/
+      _core=https://www.archlinux.org/packages/core/x86_64/$_pkgname/
+      _extra=https://www.archlinux.org/packages/extra/x86_64/$_pkgname/
+      _community=https://www.archlinux.org/packages/community/x86_64/$_pkgname/
+      for url in $_libre $_core $_extra $_community; do
+        if ! curl -s $url | grep -iq 'not found'; then
+          src=$(curl -s $url | grep -i 'source files' | cut -d'"' -f2 | sed 's#/tree/#/plain/#')
+          for link in $(curl -sL $src | grep '^  <li><a href' | cut -d"'" -f2 \
+              | sed "s#^#$(echo $src | awk -F/ '{print $3}')#"); do
+            wget -q $link -O $(basename ${link%\?*});
+          done
+          break
+        fi
+      done
 
-    # patch for cross-compiling
-    [ -f "$_srcdir"/patches/$_pkgname.patch ] || die "missing package patch"
-    cp PKGBUILD{,.old}
-    patch -Np1 -i "$_srcdir"/patches/$_pkgname.patch
+      # patch for cross-compiling
+      [ -f "$_srcdir"/patches/$_pkgname.patch ] || die "missing package patch"
+      cp PKGBUILD{,.old}
+      patch -Np1 -i "$_srcdir"/patches/$_pkgname.patch
 
-    # substitute common variables
-    sed -i "s#@TARGET@#$CHOST#g; \
-            s#@GCC_MARCH@#$GCC_MARCH#g; \
-            s#@GCC_MABI@#$GCC_MABI#g; \
-            s#@BUILDHOST@#$_buildhost#g; \
-            s#@SYSROOT@#$_sysroot#g; \
-            s#@LINUX_ARCH@#$LINUX_ARCH#g" \
-      PKGBUILD
+      # substitute common variables
+      sed -i "s#@TARGET@#$CHOST#g; \
+              s#@GCC_MARCH@#$GCC_MARCH#g; \
+              s#@GCC_MABI@#$GCC_MABI#g; \
+              s#@BUILDHOST@#$_buildhost#g; \
+              s#@SYSROOT@#$_sysroot#g; \
+              s#@LINUX_ARCH@#$LINUX_ARCH#g" \
+        PKGBUILD
 
-    # enable the target CARCH in arch array
-    sed -i "s/arch=([^)]*/& $CARCH/" PKGBUILD
+      # enable the target CARCH in arch array
+      sed -i "s/arch=([^)]*/& $CARCH/" PKGBUILD
 
-    # build the package
-    chown -R $SUDO_USER "$_makepkgdir"/$_pkgname
-    sudo -u $SUDO_USER \
-    "$_builddir"/makepkg-$CARCH.sh -fLC --config "$_builddir"/makepkg-$CARCH.conf \
-      --skipchecksums --skippgpcheck --nocheck --nodeps
+      # build the package
+      chown -R $SUDO_USER "$_makepkgdir"/$_pkgname
+      sudo -u $SUDO_USER \
+      "$_builddir"/makepkg-$CARCH.sh -fLC --config "$_builddir"/makepkg-$CARCH.conf \
+        --skipchecksums --skippgpcheck --nocheck --nodeps
 
-    # construct pkgver from pkgbuild
-    _srcinfo=$(sudo -u $SUDO_USER makepkg --printsrcinfo)
-    _pkgver=$(echo "$_srcinfo" | grep 'pkgver =' | head -n1 | awk '{print $3}')
-    _pkgrel=$(echo "$_srcinfo" | grep 'pkgrel =' | head -n1 | awk '{print $3}')
-    _epoch=$(echo "$_srcinfo" | grep 'epoch =' | head -n1 | awk '{print $3}'): || _epoch=""
-    _pkgver="$_epoch$_pkgver-$_pkgrel"
+      # construct pkgver from pkgbuild
+      _srcinfo=$(sudo -u $SUDO_USER makepkg --printsrcinfo)
+      _pkgver=$(echo "$_srcinfo" | grep 'pkgver =' | head -n1 | awk '{print $3}')
+      _pkgrel=$(echo "$_srcinfo" | grep 'pkgrel =' | head -n1 | awk '{print $3}')
+      _epoch=$(echo "$_srcinfo" | grep 'epoch =' | head -n1 | awk '{print $3}'): || _epoch=""
+      _pkgver="$_epoch$_pkgver-$_pkgrel"
+    fi
+
+    popd >/dev/null
+
+    # update pacman cache
+    rm -rf "$_chrootdir"/var/cache/pacman/pkg/*
+    rm -rf "$_pkgdest"/cross.{db,files}*
+    repo-add -q -R "$_pkgdest"/{cross.db.tar.gz,*.pkg.tar.xz}
   fi
-
-  popd >/dev/null
-
-  # update pacman cache
-  rm -rf "$_chrootdir"/var/cache/pacman/pkg/*
-  rm -rf "$_pkgdest"/repo.{db,files}*
-  repo-add -q -R "$_pkgdest"/{repo.db.tar.gz,*.pkg.tar.xz}
 
   # install in chroot
   set +o pipefail
@@ -161,6 +173,8 @@ EOF
   # remove pkg from deptree
   sed -i "/^$_pkgname :/d; s/ $_pkgname\b//g" "$_deptree"
 done
+
+echo "all packages built."
 
 # unmount sysroot
 umount "$_sysroot"/usr
